Provided access to application to use the FilePointer to read the data in a smaller chunk. (#811)
* Provided access to application to use the FilePointer to read the data in a smaller chunk.

The existing code(FilePointer) allocates buffer for the full file and returning to application for use.
If the cached file is bigger file say 20MB, it allocates for 30MB and store the data for use.
Now added the provision to application to read it however it wants. Also there is no change in existing
file reading machanics unless application sets setReadCachedDataInChunks.

Signed-off-by: Suresh Sathiah <Suresh_Sathiah2@cable.comcast.com>

* Provided access to application to use the FilePointer to read the data in a smaller chunk.

The existing code(FilePointer) allocates buffer for the full file and returning to application for use.
If the cached file is bigger file say 20MB, it allocates for 30MB and store the data for use.
Now added the provision to application to read it however it wants. Also there is no change in existing
file reading machanics unless application sets setReadCachedDataInChunks.

Signed-off-by: Suresh Sathiah <Suresh_Sathiah2@cable.comcast.com>