Fix double free error with sandbox imports
For some reason, the `appSandbox.importTracking[bPath]` is being
evaluated multiple times.  During the first pass, that whole object is
deleted, so on the second pass, it evaluates to undefined and the module
import can't finish.

I don't know exactly what is causing this problem, but this fix
re-enables things that ran under 0.35.0.x. The problem _might_ be
related to calling px.import multiple times from the same file, but
that's only a guess based on log output and what I suspect this module
might be for.