Exposes git version info to the pxscene runtime
The exposed information is derived from `git describe --tag --always
--long --dirty`, which will expose information in the form of:
`tag-offset-hash(-dirty)?`.  This is broken then loaded into the
scene.info.build variable as:
  - major
  - minor
  - micro
  - nano
  - tag
  - offset
  - hash
  - dirty

The major / minor / micro / nano are formed from the four period
separated values in the tag, or zeros if it can't be parsed.

In the case that the GIT_PXSCENE_VERSION variable isn't provided, it
will default to `0.0.0.0-0-0000000000-dirty`.

This will help developers to determine the environment that their
application has run against down to the commit.