Correct breakpad patching, simpler ucontext patch
The breakpad patching wasn't working, at the very least, because the
patch path assumed that quilt was being run from outside of the breakpad
directory.  The patch path has been corrected.

The patch that is now applied is also simpler now, as it only entails
the addition of six lines that should work whether ucontext or
ucontext_t are being provided by the glibc API.