Fix broken delListener() handling.
Now when client makes a call to delListener() everytime rtFunctionRef() is created with new
Id and server doesn't remove any handler. This change will ensure everytime rtFuncionRef is
created for rtFunctionCallback it gets the same Id. Now at the server side, the received
name (i.e. 'func://<uid>') is used to calculate a hash which is then used for add / del listners